utils: remove some control flow. add base58.
