+++ lib/db/rbt.js
@@ -15,9 +15,9 @@ var SENTINEL;
 
 /**
  * An iterative red black tree.
- * Used for the mempool. Many of its
- * options, parameters, and methods
- * mimic the leveldown interface.
+ * Many of its options, parameters,
+ * and methods mimic the leveldown
+ * interface.
  * @exports RBT
  * @constructor
  * @param {String?} location - Phony location.
@@ -586,7 +586,7 @@ RBT.prototype.open = function open(options, callback) {
 
   this.options = options;
 
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /**
@@ -595,7 +595,7 @@ RBT.prototype.open = function open(options, callback) {
  */
 
 RBT.prototype.close = function close(callback) {
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /**
@@ -622,13 +622,18 @@ RBT.prototype.get = function get(key, options, callback) {
     err = new Error('RBT_NOTFOUND: Key not found.');
     err.notFound = true;
     err.type = 'NotFoundError';
-    return utils.asyncify(callback)(err);
+    utils.nextTick(function() {
+      callback(err);
+    });
+    return;
   }
 
   if (options.asBuffer === false)
     value = value.toString('utf8');
 
-  return utils.asyncify(callback)(null, value);
+  utils.nextTick(function() {
+    callback(null, value);
+  });
 };
 
 /**
@@ -647,7 +652,7 @@ RBT.prototype.put = function put(key, value, options, callback) {
 
   this.insert(key, value);
 
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /**
@@ -665,7 +670,7 @@ RBT.prototype.del = function del(key, options, callback) {
 
   this.remove(key);
 
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /**
@@ -736,7 +741,9 @@ RBT.prototype.approximateSize = function approximateSize(start, end, callback) {
     size += item.value.length;
   }
 
-  return utils.asyncify(callback)(null, size);
+  utils.nextTick(function() {
+    callback(null, size);
+  });
 };
 
 /**
@@ -746,7 +753,7 @@ RBT.prototype.approximateSize = function approximateSize(start, end, callback) {
  */
 
 RBT.destroy = function destroy(location, callback) {
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /**
@@ -756,7 +763,7 @@ RBT.destroy = function destroy(location, callback) {
  */
 
 RBT.repair = function repair(location, callback) {
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /**
@@ -940,17 +947,28 @@ Batch.prototype.del = function del(key) {
 Batch.prototype.write = function write(callback) {
   var i, op;
 
-  if (!this.tree)
-    return callback(new Error('Already written.'));
+  if (!this.tree) {
+    utils.nextTick(function() {
+      callback(new Error('Already written.'));
+    });
+    return;
+  }
 
   for (i = 0; i < this.ops.length; i++) {
     op = this.ops[i];
-    if (op.type === 'put')
-      this.tree.insert(op.key, op.value);
-    else if (op.type === 'del')
-      this.tree.remove(op.key);
-    else
-      assert(false);
+    switch (op.type) {
+      case 'put':
+        this.tree.insert(op.key, op.value);
+        break;
+      case 'del':
+        this.tree.remove(op.key);
+        break;
+      default:
+        utils.nextTick(function() {
+          callback(new Error('Bad operation: ' + op.type));
+        });
+        return;
+    }
   }
 
   this.ops.length = 0;
@@ -1029,8 +1047,12 @@ function Iterator(tree, options) {
 Iterator.prototype.next = function(callback) {
   var item, key, value;
 
-  if (this.ended)
-    return utils.asyncify(callback)(new Error('Cannot call next after end.'));
+  if (this.ended) {
+    utils.nextTick(function() {
+      callback(new Error('Cannot call next after end.'));
+    });
+    return;
+  }
 
   if (this.options.reverse)
     item = this.snapshot[this.index--];
@@ -1040,13 +1062,15 @@ Iterator.prototype.next = function(callback) {
   if (this.options.limit != null) {
     if (this.total++ >= this.options.limit) {
       this._end();
-      return utils.nextTick(callback);
+      utils.nextTick(callback);
+      return;
     }
   }
 
   if (!item) {
     this._end();
-    return utils.nextTick(callback);
+    utils.nextTick(callback);
+    return;
   }
 
   key = item.key;
@@ -1064,7 +1088,9 @@ Iterator.prototype.next = function(callback) {
   if (this.options.valueAsBuffer === false)
     value = value.toString('utf8');
 
-  utils.asyncify(callback)(null, key, value);
+  utils.nextTick(function() {
+    callback(null, key, value);
+  });
 };
 
 /**
@@ -1105,13 +1131,17 @@ Iterator.prototype._end = function end() {
  */
 
 Iterator.prototype.end = function end(callback) {
-  if (this.ended)
-    return utils.asyncify(callback)(new Error('Already ended.'));
+  if (this.ended) {
+    utils.nextTick(function() {
+      callback(new Error('Already ended.'));
+    });
+    return;
+  }
 
   this.ended = true;
   this._end();
 
-  return utils.nextTick(callback);
+  utils.nextTick(callback);
 };
 
 /*
