+++ lib/script/script.js
@@ -18,7 +18,7 @@ var BufferWriter = require('../utils/writer');
 var BufferReader = require('../utils/reader');
 var opcodes = constants.opcodes;
 var STACK_TRUE = new Buffer([1]);
-var STACK_FALSE = new Buffer([]);
+var STACK_FALSE = new Buffer(0);
 var STACK_NEGATE = new Buffer([0x81]);
 var ScriptError = require('../utils/errors').ScriptError;
 var scriptTypes = constants.scriptTypes;
@@ -1224,7 +1224,7 @@ Script.num = function num(value, flags, size) {
  * numbers to a little-endian buffer while taking into
  * account negative zero, minimaldata, etc.
  * @example
- * assert.deepEqual(Script.array(0), new Buffer([]));
+ * assert.deepEqual(Script.array(0), new Buffer(0));
  * assert.deepEqual(Script.array(0xffee), new Buffer([0xee, 0xff, 0x00]));
  * assert.deepEqual(Script.array(new bn(0xffee)), new Buffer([0xee, 0xff, 0x00]));
  * assert.deepEqual(Script.array(new bn(0x1e).ineg()), new Buffer([0x9e]));
