+++ lib/script/witness.js
@@ -14,7 +14,7 @@ var constants = require('../protocol/constants');
 var utils = require('../utils/utils');
 var assert = require('assert');
 var opcodes = constants.opcodes;
-var STACK_FALSE = new Buffer([]);
+var STACK_FALSE = new Buffer(0);
 var STACK_NEGATE = new Buffer([0x81]);
 var scriptTypes = constants.scriptTypes;
 var Script = require('./script');
