+++ lib/utils/utils.js
@@ -16,7 +16,7 @@
 var utils = exports;
 
 var assert = require('assert');
-var native = require('./native');
+var base58 = require('./base58');
 var bn = require('bn.js');
 var util = require('util');
 var Number, Math, Date;
@@ -103,126 +103,26 @@ utils.copy = function copy(data) {
   return clone;
 };
 
-/*
- * Base58
- */
-
-var base58 = ''
-  + '123456789'
-  + 'ABCDEFGHJKLMNPQRSTUVWXYZ'
-  + 'abcdefghijkmnopqrstuvwxyz';
-
-var unbase58 = {};
-
-for (var i = 0; i < base58.length; i++)
-  unbase58[base58[i]] = i;
-
 /**
  * Encode a base58 string.
  * @see https://github.com/bitcoin/bitcoin/blob/master/src/base58.cpp
+ * @function
  * @param {Buffer} data
  * @returns {Base58String}
  */
 
-utils.toBase58 = function toBase58(data) {
-  var zeroes = 0;
-  var length = 0;
-  var str = '';
-  var i, b58, carry, j, k;
-
-  for (i = 0; i < data.length; i++) {
-    if (data[i] !== 0)
-      break;
-    zeroes++;
-  }
-
-  b58 = new Buffer(((data.length * 138 / 100) | 0) + 1);
-  b58.fill(0);
-
-  for (; i < data.length; i++) {
-    carry = data[i];
-    j = 0;
-    for (k = b58.length - 1; k >= 0; k--, j++) {
-      if (carry === 0 && j >= length)
-        break;
-      carry += 256 * b58[k];
-      b58[k] = carry % 58;
-      carry = carry / 58 | 0;
-    }
-    assert(carry === 0);
-    length = j;
-  }
-
-  i = b58.length - length;
-  while (i < b58.length && b58[i] === 0)
-    i++;
-
-  for (j = 0; j < zeroes; j++)
-    str += '1';
-
-  for (; i < b58.length; i++)
-    str += base58[b58[i]];
-
-  return str;
-};
-
-if (native)
-  utils.toBase58 = native.toBase58;
+utils.toBase58 = base58.toBase58;
 
 /**
  * Decode a base58 string.
  * @see https://github.com/bitcoin/bitcoin/blob/master/src/base58.cpp
+ * @function
  * @param {Base58String} str
  * @returns {Buffer}
  * @throws on non-base58 character.
  */
 
-utils.fromBase58 = function fromBase58(str) {
-  var zeroes = 0;
-  var i = 0;
-  var b256, ch, carry, j, out;
-
-  for (i = 0; i < str.length; i++) {
-    if (str[i] !== '1')
-      break;
-    zeroes++;
-  }
-
-  b256 = new Buffer(((str.length * 733) / 1000 | 0) + 1);
-  b256.fill(0);
-
-  for (; i < str.length; i++) {
-    ch = unbase58[str[i]];
-    if (ch == null)
-      throw new Error('Non-base58 character.');
-
-    carry = ch;
-    for (j = b256.length - 1; j >= 0; j--) {
-      carry += 58 * b256[j];
-      b256[j] = carry % 256;
-      carry = carry / 256 | 0;
-    }
-
-    assert(carry === 0);
-  }
-
-  i = 0;
-  while (i < b256.length && b256[i] === 0)
-    i++;
-
-  out = new Buffer(zeroes + (b256.length - i));
-
-  for (j = 0; j < zeroes; j++)
-    out[j] = 0;
-
-  while (i < b256.length)
-    out[j++] = b256[i++];
-
-  return out;
-};
-
-if (native)
-  utils.fromBase58 = native.fromBase58;
+utils.fromBase58 = base58.fromBase58;
 
 /**
  * Test whether a string is base58 (note that you
@@ -325,43 +225,6 @@ if (typeof setImmediate === 'function') {
     setTimeout(fn, 1);
   };
 }
-
-/**
- * Wrap a function in a `nextTick`.
- * @returns {Promise}
- * @returns {Function} Asyncified function.
- */
-
-utils.asyncify = function asyncify(callback) {
-  if (callback && callback._asyncified)
-    return callback;
-
-  function asyncifyFn(err, result1, result2) {
-    if (!callback)
-      return;
-    utils.nextTick(function() {
-      callback(err, result1, result2);
-    });
-  }
-
-  asyncifyFn._asyncified = true;
-  if (callback)
-    asyncifyFn._once = callback._once;
-
-  return asyncifyFn;
-};
-
-/**
- * Ensure a callback exists, return a NOP if not.
- * @returns {Promise}
- * @returns {Function}
- */
-
-utils.ensure = function ensure(callback) {
-  if (!callback)
-    return utils.nop;
-  return callback;
-};
 
 /**
  * Reverse a hex-string (used because of
@@ -1626,199 +1489,6 @@ utils.icmp = function icmp(target, data, start) {
 };
 
 /**
- * Asnchronously iterate over a range in parallel.
- * @param {Number} from
- * @param {Number} to
- * @param {Function} iter
- * @returns {Promise}
- */
-
-utils.forRange = function forRange(from, to, iter, callback) {
-  var pending = to - from;
-  var i, error;
-
-  callback = utils.asyncify(callback);
-
-  if (pending <= 0)
-    return callback();
-
-  function next(err) {
-    assert(pending > 0);
-    if (err)
-      error = err;
-    if (!--pending)
-      callback(error);
-  }
-
-  for (i = from; i < to; i++)
-    iter(i, next, i);
-};
-
-/**
- * Asynchronously iterate over an array in parallel.
- * @param {Array} obj
- * @param {Function} iter
- * @returns {Promise}
- */
-
-utils.forEach = function forEach(obj, iter, callback) {
-  var pending = obj.length;
-  var error;
-
-  callback = utils.asyncify(callback);
-
-  if (!pending)
-    return callback();
-
-  function next(err) {
-    assert(pending > 0);
-    if (err)
-      error = err;
-    if (!--pending)
-      callback(error);
-  }
-
-  obj.forEach(function(item, i) {
-    iter(item, next, i);
-  });
-};
-
-/**
- * Asnchronously iterate over a range in serial.
- * @param {Number} from
- * @param {Number} to
- * @param {Function} iter
- * @returns {Promise}
- */
-
-utils.forRangeSerial = function forRangeSerial(from, to, iter, callback) {
-  var called = false;
-
-  callback = utils.ensure(callback);
-
-  (function next(err) {
-    assert(!called);
-    if (err) {
-      called = true;
-      return callback(err);
-    }
-    if (from >= to) {
-      called = true;
-      return callback();
-    }
-    from++;
-    utils.nextTick(function() {
-      iter(from - 1, next, from - 1);
-    });
-  })();
-};
-
-/**
- * Asynchronously iterate over an array in serial.
- * @param {Array} obj
- * @param {Function} iter
- * @returns {Promise}
- */
-
-utils.forEachSerial = function forEachSerial(obj, iter, callback) {
-  var i = 0;
-  var called = false;
-
-  callback = utils.ensure(callback);
-
-  (function next(err) {
-    var item;
-    assert(!called);
-    if (err) {
-      called = true;
-      return callback(err);
-    }
-    if (i >= obj.length) {
-      called = true;
-      return callback();
-    }
-    item = obj[i];
-    i++;
-    utils.nextTick(function() {
-      iter(item, next, i - 1);
-    });
-  })();
-};
-
-/**
- * Asynchronously apply a truth test to every
- * member of an array in parallel.
- * @param {Array} obj
- * @param {Function} iter
- * @returns {Promise}
- */
-
-utils.every = function every(obj, iter, callback) {
-  var pending = obj.length;
-  var result = true;
-  var error;
-
-  callback = utils.asyncify(callback);
-
-  if (!pending)
-    return callback(null, result);
-
-  function next(err, res) {
-    assert(pending > 0);
-    if (err)
-      error = err;
-    if (!res)
-      result = false;
-    if (!--pending) {
-      if (error)
-        return callback(error);
-      callback(null, result);
-    }
-  }
-
-  obj.forEach(function(item, i) {
-    iter(item, next, i);
-  });
-};
-
-/**
- * Asynchronously apply a truth test to every
- * member of an array in serial.
- * @param {Array} obj
- * @param {Function} iter
- * @returns {Promise}
- */
-
-utils.everySerial = function everySerial(obj, iter, callback) {
-  var i = 0;
-  var called = false;
-
-  callback = utils.ensure(callback);
-
-  (function next(err, res) {
-    var item;
-    assert(!called);
-    if (err) {
-      called = true;
-      return callback(err);
-    }
-    if (!res) {
-      called = true;
-      return callback(null, false);
-    }
-    if (i >= obj.length) {
-      called = true;
-      return callback(null, true);
-    }
-    item = obj[i];
-    i++;
-    utils.nextTick(function() {
-      iter(item, next, i - 1);
-    });
-  })(null, true);
-};
-
-/**
  * Convert bytes to mb.
  * @param {Number} size
  * @returns {Number} mb
@@ -1862,33 +1532,6 @@ utils.inherits = function inherits(obj, from) {
   f.prototype = from.prototype;
   obj.prototype = new f;
   obj.prototype.constructor = obj;
-};
-
-/**
- * Wrap a callback to ensure it is only called once.
- * @returns {Promise}
- * @returns {Function} Wrapped callback.
- */
-
-utils.once = function once(callback) {
-  var called;
-
-  if (callback && callback._once)
-    return callback;
-
-  function onceFn(err, result1, result2) {
-    if (called)
-      return;
-    called = true;
-    if (callback)
-      callback(err, result1, result2);
-  }
-
-  onceFn._once = true;
-  if (callback)
-    onceFn._asyncified = callback._asyncified;
-
-  return onceFn;
 };
 
 /**
@@ -1980,96 +1623,6 @@ utils.hex32 = function hex32(num) {
     default:
       assert(false);
   }
-};
-
-/**
- * Wrap a callback with an `unlock` callback.
- * @see Locker
- * @returns {Promise}
- * @param {Function} unlock
- * @returns {Function} Wrapped callback.
- */
-
-utils.wrap = function wrap(callback, unlock) {
-  return function(err, res1, res2) {
-    unlock();
-    if (callback)
-      callback(err, res1, res2);
-  };
-};
-
-/**
- * Execute a stack of functions in parallel.
- * @param {Function[]} stack
- * @returns {Promise}
- */
-
-utils.parallel = function parallel(stack, callback) {
-  var pending = stack.length;
-  var error;
-  var i;
-
-  callback = utils.once(callback);
-
-  if (!pending)
-    return utils.nextTick(callback);
-
-  function next(err) {
-    assert(pending > 0);
-    if (err)
-      error = err;
-    if (!--pending)
-      callback(error);
-  }
-
-  for (i = 0; i < stack.length; i++) {
-    try {
-      // if (stack[i].length >= 2) {
-      //   stack[i](error, next);
-      //   error = null;
-      //   continue;
-      // }
-      if (error)
-        continue;
-      stack[i](next);
-    } catch (e) {
-      pending--;
-      error = e;
-    }
-  }
-};
-
-/**
- * Execute a stack of functions in serial.
- * @param {Function[]} stack
- * @returns {Promise}
- */
-
-utils.serial = function serial(stack, callback) {
-  var i = 0;
-  (function next(err) {
-    var cb = stack[i++];
-
-    if (!cb)
-      return callback(err);
-
-    // if (cb.length >= 2) {
-    //   try {
-    //     return cb(err, next);
-    //   } catch (e) {
-    //     return next(e);
-    //   }
-    // }
-
-    if (err)
-      return utils.nextTick(next.bind(null, err));
-
-    try {
-      return cb(next);
-    } catch (e) {
-      return next(e);
-    }
-  })();
 };
 
 /**
